Add a notion of physical location.
This will be useful for various things in future, such as approximating the nearest consensus service (approximate because physical distance doesn't always match network distance), visualisation, simulation, etc.

LegallyIdentifiableNode now lists a physical place for the node to be in (which might not be totally accurate one day if we have super distributed mega-nodes, but will do for now).

The config file allows the admin to specify their nearest city.

There is a city database that can map such names through to coordinates, when known.

There is support for Mercator projection.