Widen XML parameters to 64 bits
The signing tool parses values in the enclave configuration as 32-bit
integers. This patch changes it to parse them as 64-bit integers, so
that it can process larger enclaves. This only affects properties such
as MaxStackSize and MaxHeapSize, while subsequent 32-bit bookkeeping,
such as TCSPolicy, MiscSelect, and MiscMask are unchanged.

Signed-off-by: Warren He <-w@berkeley.edu>