Revert design changes to attachment demoA
The attachment demo was changed in commit dbd82705aabe020dcd2f4703b1ba9b1ed848f311 to use an in-memory
attachment. The design seemed to expect that there was shared memory between the sender and receipient
nodes, which as there isn't, the EXPECTED_HASH was always all zeroes on the recipient node, and the
demo failed (that this isn't caught by integration tests is its own issue).