Change PartyAndCertificate to an aggregate class
Change PartyAndCertificate to an aggregate class instead of a subclass of Party. This
reduces the changes compared to M11, as well as avoiding risk of accidental serialization
of a PartyAndCertificate (which may be very large) where a Party is expected.

Cleaned up initial nodes known to the identity service, in particular mock nodes now know
about themselves; previously full nodes registered themselves but mock nodes did not.