Add support for serialisation of Linked Hash Maps
Currently our advice for anyone trying to serialise a hash map as a
member of a class is to use a linked hash map instead. However, we can't
actually serialise one as it's not set as a serialisable type.

The problem is the generic collections interface we're using doesn't
recognise linked hash maps as a concept so this requires a special case
to handle properly