Send signing public key with signing request
Send signing public key so that the counterparty node knows which key
it's expected to use, and so that the requesting node can verify the
result. This is mostly useful in cases where a single counterparty may
need to sign multiple times, and therefore the signatures need to be
checked against the right keys.