Move default session ID into messaging service
Move default session ID into messaging service to clean up the API. This means it no longer shows up in a Java
class (ServiceHubInternalKt) as its only member, and is instead scoped into the service that actually uses the
value. This does result in a MessagingService.Companion object (as MessagingService is an interface), however
this is probably the most sensible solution.