Rewrite ExchangeRate to use BigDecimal
Rewrite ExchangeRate to use BigDecimal for the quantity multiplication, to ensure that there is no loss of
precision during the conversion process. The previous version using double-precision floating point maths
inherently means that the precision is not fixed, but is floating.

Also extends Amount to provide a multiplication function with rounding mode specified.