Added new static constructor to CordappLoader to be able to take arbitrary paths to load cordapps from in dev mode. CordappLoader can now find contract classes.
Minor doc changes.

Cordapp scanning is now internal and returns a list of cordapps

Cordapps now contain services.

Cordapps are now a property on the cordapp loader that are lazily loaded.

Plugins are now loaded by the cordapp loader and associated with a CorDapp

JAR path stored with CorDapp

Moved the cordapp classes to the cordapp package.

Code cleanup

Review fixes.

Review fixes.