Move to using strings in CordformContext
Move to using strings in CordformContext; X500Name is an external API dependency we're trying to avoid,
X500Principal rearranges the order of the elements in a way we don't want, and trying to put CordaX500Name
into Groovy in the time available is impractical. As such having pre-formatted strings used in the Cordform
plugin makes most sense right now.