Move to a CSR-like format for identity assertion
Move away from using challenge-response nonce approach (with the
problems associated with the counterparty defining the data to sign),
and to a CSR-like format. The data blob is considered an assertion
that the signing key's holder wants to use the key to represent the
specified X.500 name, and is then signed by the private key.

Instead of a conventional CSR this wraps the CSR info in fixed
header/footer so it can't be reused in different contexts.