Use Java reflection for determining if a type is a Kotlin singleton.
Kotlin's own reflection has a habit of throwing weird errors and
does not work for private objects - thus the unit test which tests
this feature has actually never worked properly, but somehow works
by accident. An attempt to upgrade to the latest Kotlin revealed
the issue so it must have always been unstable.