CORDA-1130 - Round trip serialization of the TVE exceptions fail
This appears to be because the way they're written simply makes them
not serializable by the AMQP library. They could be more simply fixed to
map the missing parameters to properties, but then the round trip
serialization will fail as the constructors have side effects

The only way to get around this *and* to preserve the clean interface to
those calling the exceptions from user code is to allow the serializer
to access the main constructor and use a secondary for external access.

To disambiguate the signatures where needed use a dummy parameter that is
understood by the Serializer and that it know to not include in the
serialized form. However, to enable this, this does mean adding a new
constructor method to Flow Exception

Also create a new module for common elements that node-api and core need
to access, like the serialization exception.