ENT-1665 - Upgrade fast-classpath-scanner library to 2.12.3
Specifically, pulls in a fix we're seeing on systems when Corda
is deployed as part of Spring Boot where the scanner fails.

Currently can't switch to the latest version, 2.18.1, as that exposes
a bug in the classpath scanner when we try building our docs. (see
ENT-1665 for a description of this). For now, lets just go with
a newer version of the library we know works and worry about moving
to the absolute latest version at a later date