Draft to add more node schema related table info
Document the nature of each of the node related standard tables in the database.  Which may also include in the future recommendations to node users of how to housekeep and maintain these tables ... and other ways to correlate data found in the database, with information represented in the log files.