ENT-2489: Additional optional parameter to specify target for RPC invocation.
This is a split-out change from an enterprise feature, but since we do have: `RPCApi.RPC_TARGET_LEGAL_IDENTITY` header constant even in OS, it would make sense to extend RPC client API to populate this.

Also change to `SessionRejectException` without which it will not be AMQP serializable and will error at serialization time.