Restore separate enterprise implementation of cash selection for SQL server (#1421)
Create Enterprise version of CashSelectionSQLServerImpl and copy the content of CashSelectionSQLServerImpl as of 0c8c914 commit (with relevancy_status added and no copyrights).
The commit a28767343222ab5b7afdb1b546f6204fcaa69103 (OS--> ENT merge) replaced the original enterprise code an community independent version (they had filename in common only). To avoid this clash the original enterprise implementation will be now in EnterpriseCashSelectionSQLServerImpl (and is the default one for ServiceLoader) and  CashSelectionSQLServerImpl remains an empty placeholder in Enterprise.